+++ Server/RequestEnvironment.mjs
@@ -24,6 +24,8 @@ class RequestEnvironment {
 		this.cookie = cookie.parse(request.headers["cookie"] || "");
 		/**
 		 * The parsed post data as an object, if applicable.
+		 * 
+		 * Note: This is not prefilled automatically. You need to attach some middleware for this to work. Since v1.3.0 powahroot comes with some basic default middleware such as `middleware_parse_json` populates this variable.
 		 * @type {Object|null}
 		 */
 		this.post_data = null;