+++ Server/middleware/middleware.mjs
@@ -1,0 +1,12 @@
+"use strict";
+
+import middleware_log_requests from './middleware_log_requests.mjs';
+import middleware_catch_errors from './middleware_catch_errors.mjs';
+import middleware_parse_json from './middleware_parse_json.mjs';
+
+// NOTE: Don't forget to update Server.mjs too!!
+export {
+	middleware_log_requests,
+	middleware_catch_errors,
+	middleware_parse_json
+};